backend/requirements.txt
ansicon==1.89.0
asgiref==3.7.2
awsebcli==3.23.3
blessed==1.21.0
boto3==1.38.5
botocore==1.38.5
cement==2.10.14
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
clr_loader==0.2.7.post0
colorama==0.4.6
decorator==5.2.1
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.3
django-storages==1.14.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
ffmpeg-python==0.2.0
future==1.0.0
gunicorn==23.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflection==0.5.1
jinxed==1.3.0
jmespath==1.0.1
moviepy==2.1.2
msgpack==1.1.0
mutagen==1.47.0
numpy==2.2.5
packaging==24.2
pathspec==0.12.1
pillow==10.4.0
proglog==0.1.11
pycparser==2.22
pydub==0.25.1
PyJWT==2.10.1
PyMySQL==1.1.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pythonnet==3.0.5
pytz==2024.1
pywin32==310
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
s3transfer==0.12.0
semantic-version==2.10.0
setuptools==80.3.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
tqdm==4.67.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
wcwidth==0.2.13
